+++ packages/react-magma-dom/src/components/Select/Select.stories.tsx
@@ -80,7 +80,6 @@ export const Default = {
     isLabelVisuallyHidden: false,
     isMulti: false,
     labelPosition: LabelPosition.top,
-    ariaLabel: 'Another select text',
   },
 };
 
@@ -108,7 +107,6 @@ export const Multi = {
   args: {
     ...Default.args,
     disabled: false,
-    ariaLabel: 'Multi select example',
   },
 };