+++ packages/react-magma-dom/src/components/Textarea/index.tsx
@@ -76,11 +76,18 @@ export const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
     const theme = React.useContext(ThemeContext);
 
     const id = useGenerateId(defaultId);
-    const descriptionId =
-      errorMessage || helperMessage || maxCount || maxLength
-        ? `${id}__desc`
-        : null;
+
     const maxCharacters = typeof maxCount === 'number' ? maxCount : maxLength;
+
+    const counterDescriptionId =
+      typeof maxCharacters === 'number' && hasCharacterCounter
+        ? `${id}__counter`
+        : null;
+    const messageDescriptionId =
+      errorMessage || helperMessage ? `${id}__message` : null;
+    const descriptionId =
+      [counterDescriptionId, messageDescriptionId].filter(Boolean).join(' ') ||
+      null;
 
     const maxLengthNum =
       !hasCharacterCounter && maxLength ? maxLength : undefined;
@@ -128,7 +135,7 @@ export const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
       >
         <StyledTextArea
           {...other}
-          aria-labelledby={
+          aria-describedby={
             descriptionId ? descriptionId : props['aria-describedby']
           }
           aria-invalid={!!errorMessage}